+++ django/applications/catmaid/static/libs/jquery/jquery.tag.editor.js
@@ -185,6 +185,7 @@ Joost Elfering
       {
         tagEditorGetTags: function ()
         {
+          buildArray();
           return itemBase.join(options.separator);
         },
         tagEditorResetTags: function ()