+++ django/applications/catmaid/urls.py
@@ -305,6 +305,7 @@ urlpatterns += [
 urlpatterns += [
     url(r'^dataviews/list$', data_view.get_available_data_views, name='list_dataviews'),
     url(r'^dataviews/default$', data_view.get_default_properties, name='default_dataview'),
+    url(r'^dataviews/(?P<data_view_id>\d+)/$', data_view.get_detail, name='detail_dataview'),
     url(r'^dataviews/show/(?P<data_view_id>\d+)$', data_view.get_data_view, name='show_dataview'),
     url(r'^dataviews/show/default$', data_view.get_default_data_view, name='show_default_dataview'),
     url(r'^dataviews/type/comment$', data_view.get_data_view_type_comment, name='get_dataview_type_comment'),
@@ -407,6 +408,13 @@ urlpatterns += [
    url(r'^(?P<project_id>\d+)/volumes/(?P<volume_id>\d+)/intersect$', volume.intersects),
 ]
 
+# Analytics
+UrlParser.explicit_root_paths |= set(['{project_id}/analytics'])
+urlpatterns += [
+    url(r'^(?P<project_id>\d+)/analytics/skeletons$', analytics.analyze_skeletons),
+    url(r'^(?P<project_id>\d+)/analytics/broken-section-nodes$', analytics.list_broken_section_nodes)
+]
+
 # Front-end tests
 urlpatterns += [
     url(r'^tests$', login_required(CatmaidView.as_view(template_name="catmaid/tests.html")), name="frontend_tests"),
@@ -428,9 +436,6 @@ urlpatterns += [
     # Circles
     url(r'^(?P<project_id>\d+)/graph/circlesofhell', circles.circles_of_hell),
     url(r'^(?P<project_id>\d+)/graph/directedpaths', circles.find_directed_paths),
-
-    # Analytics
-    url(r'^(?P<project_id>\d+)/skeleton/analytics$', analytics.analyze_skeletons),
 
     # Review
     url(r'^(?P<project_id>\d+)/user/reviewer-whitelist$', review.reviewer_whitelist),